refactor: move component theme shell to custom
